Renamed main to p2sm
