+++ .gitignore
@@ -34,3 +34,4 @@
 p2sm.vcxproj.user
 /.idea
 p2sm.sln.DotSettings.user
+copy.bat
